合并两个工具文件
