--- packages/fiber/util.js
@@ -1,12 +1,0 @@
-export const gSBU = "getSnapshotBeforeUpdate";
-export const gDSFP = "getDerivedStateFromProps";
-
-export let effects = [];
-export function resetStack(info) {
-    keepLast(info.containerStack);
-    keepLast(info.containerStack);
-}
-function keepLast(list) {
-    var n = list.length;
-    list.splice(0, n - 1);
-}
